Clarify purpose of repo in README.md (#14)
I hope my assumption is correct: That this repo exists to demonstrate the usage of your [`terraform-aws-serverless`](https://github.com/FormidableLabs/terraform-aws-serverless#reference-project), as hinted in the docs for that module:

> Perhaps the easiest place to start is our sample reference project that creates a Serverless framework service named simple-reference that integrates the core module and submodules of this project

---

If that is true then it can be a little confusing landing on this reference repo (e.g. from Google), and it not being explicit about its purpose.